test: extract suite
